+++ tests/rules/test_duplicates.py
@@ -1,81 +1,53 @@
 import arche.rules.duplicates as duplicates
-from arche.rules.result import Level, Outcome
+from arche.rules.result import Level
 from conftest import *
 import numpy as np
 import pandas as pd
 import pytest
 
 
-unique_inputs = [
-    ({}, {}, {Level.INFO: [(Outcome.SKIPPED,)]}),
-    (
-        {"id": ["0", "0", "1"]},
-        {"unique": ["id"]},
-        {
-            Level.ERROR: [
-                ("id contains 1 duplicated value(s)", None, {"same '0' `id`": [0, 1]})
-            ]
-        },
-    ),
-    (
-        {
-            "id": ["47" for x in range(6)],
-            "name": ["Walt", "Juan", "Juan", "Walt", "Walt", "John"],
-        },
-        {"unique": ["id", "name"]},
-        {
-            Level.ERROR: [
-                (
-                    "id contains 1 duplicated value(s)",
-                    None,
-                    {"same '47' `id`": [i for i in range(6)]},
-                ),
-                (
-                    "name contains 2 duplicated value(s)",
-                    None,
-                    {"same 'Juan' `name`": [1, 2], "same 'Walt' `name`": [0, 3, 4]},
-                ),
-            ]
-        },
-    ),
-    ({"name": ["a", "b"]}, {"unique": ["name"]}, {}),
-]
-
-
-@pytest.mark.parametrize("data, tagged_fields, expected_messages", unique_inputs)
-def test_find_by_unique(data, tagged_fields, expected_messages):
-    df = pd.DataFrame(data)
-    assert_results_equal(
-        duplicates.find_by_unique(df, tagged_fields),
-        create_result(
-            "Duplicates By **unique** Tag", expected_messages, items_count=len(df)
-        ),
-    )
-
-
 @pytest.mark.parametrize(
     "data, columns, expected_messages",
     [
+        ({"id": ["0", "1", "2"]}, ["id"], {}),
         (
             {"id": ["0", "0", "1"]},
             ["id"],
             {
                 Level.ERROR: [
-                    ("2 duplicate(s) with same id", None, {"same '0' `id`": [0, 1]})
+                    (
+                        "id contains 1 duplicated value(s)",
+                        None,
+                        {"same '0' `id`": [0, 1]},
+                    )
                 ]
             },
         ),
-        ({"id": ["0", "1", "2"]}, ["id"], {}),
         (
-            {"id": [np.nan, "9", "9"], "city": [np.nan, "Talca", "Talca"]},
-            ["id", "city"],
+            {
+                "id": [np.nan, "9", "9"],
+                "city": [np.nan, "Talca", "Talca"],
+                "id2": ["47" for x in range(3)],
+                "name": ["Walt", "Juan", "Juan"],
+            },
+            [["id", "city"], "id2", "name"],
             {
                 Level.ERROR: [
                     (
-                        "2 duplicate(s) with same id, city",
+                        "id, city contains 1 duplicated value(s)",
                         None,
                         {"same '9' `id`, 'Talca' `city`": [1, 2]},
-                    )
+                    ),
+                    (
+                        "id2 contains 1 duplicated value(s)",
+                        None,
+                        {"same '47' `id2`": [0, 1, 2]},
+                    ),
+                    (
+                        "name contains 1 duplicated value(s)",
+                        None,
+                        {"same 'Juan' `name`": [1, 2]},
+                    ),
                 ]
             },
         ),
@@ -92,17 +64,25 @@ def test_find_by(data, columns, expected_messages):
 @pytest.mark.parametrize(
     "data, tagged_fields, expected_messages",
     [
-        ({}, {}, {Level.INFO: [(Outcome.SKIPPED,)]}),
         (
-            {"name": ["bob", "bob", "bob", "bob"], "url": ["u1", "u1", "2", "u1"]},
-            {"name_field": ["name"], "product_url_field": ["url"]},
+            {
+                "name": ["bob", "bob", "bob", "bob"],
+                "url": ["u1", "u1", "2", "u1"],
+                "id": [np.nan, "9", "9", None],
+            },
+            {"name_field": ["name"], "product_url_field": ["url"], "unique": ["id"]},
             {
                 Level.ERROR: [
                     (
-                        "3 duplicate(s) with same name, url",
+                        "id contains 1 duplicated value(s)",
+                        None,
+                        {"same '9' `id`": [1, 2]},
+                    ),
+                    (
+                        "name, url contains 1 duplicated value(s)",
                         None,
                         {"same 'bob' `name`, 'u1' `url`": [0, 1, 3]},
-                    )
+                    ),
                 ]
             },
         ),
@@ -113,13 +93,9 @@ def test_find_by(data, columns, expected_messages):
         ),
     ],
 )
-def test_find_by_name_url(data, tagged_fields, expected_messages):
+def test_find_by_tags(data, tagged_fields, expected_messages):
     df = pd.DataFrame(data)
     assert_results_equal(
-        duplicates.find_by_name_url(df, tagged_fields),
-        create_result(
-            "Duplicates By **name_field, product_url_field** Tags",
-            expected_messages,
-            items_count=len(df),
-        ),
+        duplicates.find_by_tags(df, tagged_fields),
+        create_result("Duplicates", expected_messages, items_count=len(df)),
     )