+++ tests/rules/test_others.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 from arche.rules.others import compare_boolean_fields, garbage_symbols
-from arche.rules.result import Level, Outcome
+from arche.rules.result import Level
 from conftest import *
 import pandas as pd
 import pytest
@@ -19,19 +19,9 @@ compare_bool_data = [
             )
         ],
     ),
-    (
-        {"bool_f": [True], "bool_f2": [False]},
-        {"diff_bool_field": [False]},
-        {Level.INFO: [(Outcome.SKIPPED,)]},
-        [],
-    ),
-    (
-        {"bool_f": [True]},
-        {"str_f": ["True", "True"]},
-        {Level.INFO: [(Outcome.SKIPPED,)]},
-        [],
-    ),
-    ({"str_f": ["True"]}, {"bool_f": [True]}, {Level.INFO: [(Outcome.SKIPPED,)]}, []),
+    ({"bool_f": [True], "bool_f2": [False]}, {"diff_bool_field": [False]}, {}, []),
+    ({"bool_f": [True]}, {"str_f": ["True", "True"]}, {}, []),
+    ({"str_f": ["True"]}, {"bool_f": [True]}, {}, []),
     (
         {"b": [True, True, True]},
         {"b": [True]},